requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.10
django-crispy-forms==1.13.0
djangorestframework==3.13.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
oauthlib==3.1.1
Pillow==9.0.0
pkg_resources==0.0.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
